fix: 回転方向反転

+++ srcs/vector_utils.c
@@ -6,7 +6,7 @@
 /*   By: minabe <minabe@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/09 18:47:14 by minabe            #+#    #+#             */
-/*   Updated: 2023/09/12 14:57:31 by minabe           ###   ########.fr       */
+/*   Updated: 2023/09/12 15:58:19 by minabe           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ void	set_vector(t_vector *vector, double x, double y)
 	vector->y = y;
 }
 
-int	vectorlen(t_vector vector)
+double	vectorlen(t_vector vector)
 {
 	return (sqrt(vector.x * vector.x + vector.y * vector.y));
 }
